Simplifications and making fake splits consistent
For each team, they may only have 1 fake unit on the board. All other subs after being split will be real. The enemy has to  shoot the initial sub to do the standard of both removing the decoy and hitting the actual sub. The user also now if there is only one of their units on the tile, the game understands they mean to click on it and skips the unittype selection phase. If there are two or more though, it will still do so.